refactor: clean up, reorganize
