+++ SwiftDiffusion/SwiftDiffusionApp.swift
@@ -5,7 +5,6 @@
 //  Created by Justin Bush on 2/3/24.
 //
 
-import Cocoa
 import SwiftUI
 
 @main
@@ -54,44 +53,4 @@ struct SwiftDiffusionApp: App {
       }
     }
   }
-}
-
-// TODO: UI indicators for user on error
-/// Initialize app file-folder structure setup with error handling.
-func setupAppFileStructure() {
-  AppFileStructure.setup { error, failedUrl in
-    if let error = error, let failedUrl = failedUrl {
-      Debug.log("Failed to create directory at \(failedUrl): \(error)")
-    } else if let error = error {
-      Debug.log("Error: \(error)")
-    } else {
-      Debug.log("Successfully initialized application file structure.")
-    }
-  }
-  AppFileStructure.setupDocuments { error, failedUrl in
-    if let error = error, let failedUrl = failedUrl {
-      Debug.log("Failed to create directory at \(failedUrl): \(error)")
-    } else if let error = error {
-      Debug.log("Error: \(error)")
-    } else {
-      Debug.log("Successfully initialized application documents structure.")
-    }
-  }
-}
-
-
-// MARK: - AppDelegate
-class AppDelegate: NSObject, NSApplicationDelegate {
-  
-  func applicationShouldTerminate(_ sender: NSApplication) -> NSApplication.TerminateReply {
-    ScriptManager.shared.terminateImmediately()
-    
-    return .terminateNow
-  }
-  
-  
-  func applicationDidFinishLaunching(_ notification: Notification) {
-    Debug.log("applicationDidFinishLaunching")
-  }
-  
 }
